HARP-7221: Move OrientedBox3 to @here/harp-geoutils
This change moves the concrete implementation of the OrientedBox3Like
interface from @here/harp-geometry to @here/harp-geoutils. To ensure
source compatibilty a typealias to OrientedBox3 (marked as deprecated)
is added to @here/harp-geometry.
